+++ src/rendering/walls.c
@@ -6,7 +6,7 @@
 /*   By: gbuczyns <gbuczyns@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/10 18:31:28 by gbuczyns          #+#    #+#             */
-/*   Updated: 2024/12/03 17:05:28 by gbuczyns         ###   ########.fr       */
+/*   Updated: 2024/12/03 19:34:47 by gbuczyns         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,23 +15,18 @@
 #define HORIZONTAL 0
 #define VERTICAL 1
 
-
-
 typedef struct s_wall
 {
 	t_dvector	pos;
 	int			height;
-	double	perp_wall_dist;
+	double		perp_wall_dist;
 	int			y_start;
 	int			wall_end;
 	int			wall_y;
 	t_dvector	factor;
 
 	// double	tex_y;
-}			t_wall;
-
-
-
+}				t_wall;
 
 void	calculate_wall_x(t_game *game, t_wall *wall)
 {
@@ -48,8 +43,7 @@ void	calculate_wall_x(t_game *game, t_wall *wall)
 		wall->factor.x = 1 - wall->factor.x;
 }
 
-
-void calculate_wall_distance(t_game *game, t_wall *wall)
+void	calculate_wall_distance(t_game *game, t_wall *wall)
 {
 	if (game->rc.side == 0)
 		wall->perp_wall_dist = (game->rc.map.x - game->player.pos.y + (1
@@ -59,29 +53,28 @@ void calculate_wall_distance(t_game *game, t_wall *wall)
 					- game->rc.step.x) / 2) / game->rc.raydir.x + 0.0001;
 }
 
-void calculate_wall_height(t_game *game, t_wall *wall)
+void	calculate_wall_height(t_game *game, t_wall *wall)
 {
 	wall->height = (int)(SCREEN_HEIGHT / wall->perp_wall_dist);
 	wall->y_start = (SCREEN_HEIGHT - wall->height) / 2;
 	wall->wall_end = SCREEN_HEIGHT / 2 + wall->height / 2;
 }
 
-void calculate_draw_limits(t_game *game, t_wall *wall)
-{	
+void	calculate_draw_limits(t_game *game, t_wall *wall)
+{
 	game->rc.draw_start = wall->y_start;
 	if (game->rc.draw_start < 0)
 		game->rc.draw_start = 0;
 	game->rc.draw_end = wall->wall_end;
 	if (game->rc.draw_end > SCREEN_HEIGHT)
-		game->rc.draw_end = SCREEN_HEIGHT -1;
-	
+		game->rc.draw_end = SCREEN_HEIGHT - 1;
 }
 
 // void	calculate_texture_coordinates(t_game *game, t_wall *wall)
 // {
 
-
-// 	// game->rc.step_size = 1.0 * game->rc.texture->height / game->rc.line_height;
+// 	// game->rc.step_size = 1.0 * game->rc.texture->height
+		// / game->rc.line_height;
 // 	// game->rc.tex_pos = (game->rc.draw_start - SCREEN_HEIGHT / 2
 // 	// 		+ game->rc.line_height / 2) * game->rc.step_size;
 // }
@@ -89,10 +82,9 @@ void calculate_draw_limits(t_game *game, t_wall *wall)
 void	print_stripes(t_game *game)
 {
 	int				x;
-	int 			y;
+	int				y;
 	unsigned int	color;
 	t_wall			wall;
-
 
 	x = -1;
 	while (++x < SCREEN_WIDTH)
@@ -108,14 +100,15 @@ void	print_stripes(t_game *game)
 		y = game->rc.draw_start - 1;
 		while (++y < game->rc.draw_end)
 		{
-			wall.factor.y = (y - wall.y_start) / wall.height;
-			color = get_texture_pixel(game->rc.tex_img, wall.factor.y, wall.factor.x);
+			wall.factor.y = ((double)y - wall.y_start) / wall.height;
+			color = get_texture_pixel(game->rc.tex_img, wall.factor.y,
+					wall.factor.x);
 			put_pixel(&game->display, x, y, color);
 		}
 	}
 }
 
-			// render_walls(game, x, y);
+// render_walls(game, x, y);
 // if (game->rc.draw_start < 0)
 // 	wall_start = - game->rc.draw_start;
 // else
@@ -146,7 +139,3 @@ void	set_ray_steps(t_game *game)
 			game->rc.hit = 1;
 	}
 }
-
-
-
-
